requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
numpy>=1.22.0
pandas>=1.5.0
scipy>=1.9.0
matplotlib>=3.5.0
chess>=1.9.0
python-chess>=1.9.0
tqdm>=4.62.0
pytest>=7.0.0
tensorboard>=2.10.0
scikit-learn>=1.0.0
networkx>=2.7.0
pyglet>=2.0.0  # for visualization
pyzmq>=24.0.0  # for RPC
protobuf>=3.19.0  # for gRPC
grpcio>=1.50.0
pyarrow>=10.0.0  # for data storage
pyyaml>=6.0.0
rich>=12.0.0  # for fancy CLI
gymnasium>=0.28.0  # for RL environments 